requirements.txt
pytest>=8.3.3
pytest-cov>=5.0.0
pylint>=3.2.7
black>=24.8.0
transformers==4.44.2
datasets==3.0.0
haystack-ai==2.5.1
pdfminer.six==20240706
pypdf==4.3.1
markdown-it-py==3.0.0
mdit_plain==1.0.1
python-docx==1.1.2
python-pptx==1.0.2
streamlit==1.38.0
accelerate>=0.34.2
sentence-transformers>=3.0.0